simple_scheduler: rename to scheduler
